refactor: move query circuit into circuit module
